+++ scripts/cli/src/fully_managed_connector_show_config_command.sh
@@ -31,29 +31,37 @@ do
         log "🐞 curl command used"
         echo "curl $security -s -X GET -H "Content-Type: application/json" "https://api.confluent.cloud/connect/v1/environments/$environment/clusters/$cluster/connectors/$connector/config" --header "authorization: Basic $authorization""
     fi
-    json_config=$(curl $security -s -X GET -H "Content-Type: application/json" "https://api.confluent.cloud/connect/v1/environments/$environment/clusters/$cluster/connectors/$connector/config" --header "authorization: Basic $authorization")
-    echo "playground fully-managed-connector create-or-update --connector $connector << EOF"
-    echo "$json_config" | jq -S . | sed 's/\$/\\$/g'
-    echo "EOF"
-
-    if [[ "$OSTYPE" == "darwin"* ]]
+    curl_output=$(curl $security -s -X GET -H "Content-Type: application/json" "https://api.confluent.cloud/connect/v1/environments/$environment/clusters/$cluster/connectors/$connector/config" --header "authorization: Basic $authorization")
+    ret=$?
+    set -e
+    if [ $ret -eq 0 ]
     then
-        clipboard=$(playground config get clipboard)
-        if [ "$clipboard" == "" ]
-        then
-            playground config set clipboard true
-        fi
+        echo "playground fully-managed-connector create-or-update --connector $connector << EOF"
+        echo "$curl_output" | jq -S . | sed 's/\$/\\$/g'
+        echo "EOF"
 
-        if [ "$clipboard" == "true" ] || [ "$clipboard" == "" ]
+        if [[ "$OSTYPE" == "darwin"* ]]
         then
-            tmp_dir=$(mktemp -d -t ci-XXXXXXXXXX)
-            trap 'rm -rf $tmp_dir' EXIT
-            echo "playground fully-managed-connector --connector $connector << EOF" > $tmp_dir/tmp
-            echo "$json_config" | jq -S . | sed 's/\$/\\$/g' >> $tmp_dir/tmp
-            echo "EOF" >> $tmp_dir/tmp
+            clipboard=$(playground config get clipboard)
+            if [ "$clipboard" == "" ]
+            then
+                playground config set clipboard true
+            fi
 
-            cat $tmp_dir/tmp | pbcopy
-            log "📋 connector config has been copied to the clipboard (disable with 'playground config set clipboard false')"
+            if [ "$clipboard" == "true" ] || [ "$clipboard" == "" ]
+            then
+                tmp_dir=$(mktemp -d -t ci-XXXXXXXXXX)
+                trap 'rm -rf $tmp_dir' EXIT
+                echo "playground fully-managed-connector --connector $connector << EOF" > $tmp_dir/tmp
+                echo "$curl_output" | jq -S . | sed 's/\$/\\$/g' >> $tmp_dir/tmp
+                echo "EOF" >> $tmp_dir/tmp
+
+                cat $tmp_dir/tmp | pbcopy
+                log "📋 connector config has been copied to the clipboard (disable with 'playground config set clipboard false')"
+            fi
         fi
+    else
+        logerror "❌ curl request failed with error code $ret!"
+        exit 1
     fi
 done
